requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytz==2023.3



